민상훈_240116_01/회원탈퇴기능 작성중

+++ reactRepo/pupple/src/component/mypage/MemberQuit.jsx
@@ -5,7 +5,11 @@ import styled from 'styled-components';
 const StyledMemberQuitDiv = styled.div`
 
 `;
-
+const ErrorMsg = styled.span`
+  color: red;
+  display: block; // or inline-block, depending on preference
+  margin-top: 5px;
+`;
 
 const MemberQuit = () => {
     const navigate = useNavigate();
@@ -16,10 +20,17 @@ const MemberQuit = () => {
     const loginMemberVo = JSON.parse(sessionStorage.getItem("loginMemberVo"));
     console.log(loginMemberVo);
 
-    const [vo, setVo] = useState();
+    const [vo, setVo] = useState({
+        id:loginMemberVo.id, pwd:'', pwd2:''
 
+    });
+    
     const handleInputChange = (e) => {
         const {name, value} = e.target;
+        if(name === 'pwd2'){
+            setPasswordMatchError(value !== vo.pwd);
+        }
+        
         setVo({
             ...vo,
             [name]: value
@@ -35,17 +46,42 @@ const MemberQuit = () => {
         }
 
         setIsFetching(true);
+        if(vo.pwd !== vo.pwd2){
+            setPasswordMatchError(true);
+            setIsFetching(false);
+            return;
+        }
 
-
-        fetch("http://127.0.0.1:8080/app/member/quit", {
+        fetch("http://127.0.0.1:8080/app/member/mypage/memberQuit", {
             method: "DELETE",
-            headers: {"Content-Type": "application/json"
+            headers: {
+                "Content-Type": "application/json"
         },
-        body: JSON.stringify
+            body: JSON.stringify(vo)
         })
         .then( resp => {resp.json()})
-        .then( navigate("/"))
+        .then((data) => {
+            if(data.msg === "good"){
+                alert("회원 탈퇴 성공");
+                navigate("/");
+            }else{
+                alert("회원 탈퇴 실패")
+                navigate("/quit");
+            }
+        })
+        .finally( () => {
+            isFetching = false;
+        })
+        ;
     }
+    const handleComparePwd = () => {
+        if (vo.pwd !== vo.pwd2) {
+          setPasswordMatchError(true);
+        } else {
+          setPasswordMatchError(false);
+        }
+      };
+    
     return (
         <StyledMemberQuitDiv>
             <form onSubmit={handleQuitSubmit}>
@@ -65,11 +101,12 @@ const MemberQuit = () => {
                         </tr>
                         <tr>
                             <td>비밀번호</td>
-                            <td><input type="text" name="pwd" onChange={handleInputChange}/></td>
+                            <td><input type="password" name="pwd" onChange={handleInputChange}/></td>
                         </tr>
                         <tr>
                             <td>비밀번호 확인</td>
-                            <td><input type="text" name="pwd2" onChange={handleInputChange}/></td>
+                            <td><input type="password" name="pwd2" onChange={handleInputChange} onBlur={handleComparePwd}/>
+                            {passwordMatchError && <ErrorMsg>비밀번호가 일치하지 않습니다.</ErrorMsg>}</td>
                         </tr>
                         <tr>
                             <td>생년월일</td>
